requirements.txt
flask
flask-cors
numpy
Pillow
opencv-python
easyocr
